// packages/solver/src/@generic/enums/solution-technique.enum.ts
export enum SolutionTechniqueEnum {
    Guess = 0,
    FullHouse = 1,
    NakedSingle = 2,
    HiddenSingle = 3,
    PointingPair = 4,
    PointingTriple = 5,
    BoxLineReduction = 6,
    NakedPair = 7,
    NakedTriple = 8,
    NakedQuad = 9,
    HiddenPair = 10,
    HiddenTriple = 11,
    HiddenQuad = 12,
    XWing = 13,
    Swordfish = 14,
    Jellyfish = 15,
    FinnedXWing = 16,
    FinnedSwordfish,
    SashimiXWing,
    SashimiSwordfish,
    XYWing,
    XYZWing,
    WWing,
    XChain,
    XYChain,
    SimpleColoring,
    AIC
}